@@ -23,7 +23,7 @@ authors = [
     {name = "Sebastian Pölsterl", email = "sebp@k-d-w.org"}
 ]
 license = {file = "COPYING"}
-requires-python = ">=3.9"
+requires-python = ">=3.10"
 classifiers = [
     "Development Status :: 4 - Beta",
     "Intended Audience :: Science/Research",
@@ -36,10 +36,10 @@ classifiers = [
     "Programming Language :: Cython",
     "Programming Language :: Python",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
     "Programming Language :: Python :: 3.12",
+    "Programming Language :: Python :: 3.13",
     "Topic :: Software Development",
     "Topic :: Scientific/Engineering"
 ]
@@ -114,7 +114,7 @@ namespaces = false
 [tool.black]
 line-length = 120
 extend-exclude = "sksurv/linear_model/src/eigen"
-target-version = ["py39"]
+target-version = ["py310"]
 
 [tool.pytest.ini_options]
 minversion = "6.0"
@@ -183,7 +183,7 @@ extend-exclude = [
 # Group violations by containing file.
 output-format = "grouped"
 line-length = 120
-target-version = "py39"
+target-version = "py310"
 
 [tool.ruff.lint]
 ignore = ["C408"]
